refactor(gateway): extract session and cors helpers
